chore: collapsable card style props
