+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/index.tsx
@@ -4,10 +4,9 @@ import React from 'react';
 import {
     COLLAPSABLE_CARD_BUTTON_PROPS,
     COLLAPSABLE_CARD_POPOVER_PROPS,
-} from '../../common/CollapsableCard';
+} from '../../../constants';
 import MantineIcon from '../../common/MantineIcon';
 import { useVisualizationContext } from '../../LightdashVisualization/VisualizationProvider';
-
 import PieChartConfigTabs from './PieChartConfigTabs';
 
 const PieChartConfig: React.FC = () => {
